python_version/requirements.txt
chainlit==0.7.700
openai==1.12.0
python-dotenv==1.0.1 